// moonbowcl/src/app/model/dto/attr/AttrCrmConfNextCrm.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class AttrCrmConfNextCrm {
    title: string;
    group_id: string;
    priority_id: number;
    note: string;
    type_id: number;
    asset_id: string;
    additional_info: string;
    article_istanza_subject: string;
    article_istanza_body: string;
    include_allegati: boolean;
    article_allegato_subject: string;
    codiceFiscale: string;
    firstname: string;
    lastname: string;
    email: string;
    phone: string;
    mobile: string;
}